Implement Forward and Reverse traits
This is an implementation of forward and reverse geocoding traits, and
the provision of these traits for the OpenCage geocoding service.

The traits are deliberately split, and are deliberately minimal; any
geocoding provider should be able to return an address, and accept an
address, returning one or more points.

The initial implementation makes this trait available for the OpenCage
provider, and also makes its full API responses available so that
more flexible methods can be added (that isn't WIP, but they'll land
at some point in the near future)